use livepage404 but relocate it, delete internal404
